#10072: Merge atanh_bw to TTNN
