+++ docs/source/ttnn/ttnn/ttnn/atanh_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.atanh_bw:
+
+ttnn.atanh_bw
+##############
+
+  .. autofunction:: ttnn.atanh_bw
