+++ tests/tt_eager/python_api_testing/unit_testing/backward_ops/test_backward_atanh.py
@@ -4,7 +4,7 @@
 
 import torch
 import pytest
-import tt_lib
+import ttnn
 from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import (
     data_gen_with_range,
     compare_pcc,
@@ -25,7 +25,7 @@ def test_bw_atanh(input_shapes, device):
 
     pyt_y = torch.atanh(in_data)
 
-    tt_output_tensor_on_device = tt_lib.tensor.atanh_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.atanh_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
